Implement an API for forking and joining processes.
Creates a new internal AsyncProcessOperations API for creating the
forked processes and waiting for their completion. This pulls some of
the ExecHandle internal API into a public ProcessHandle API. It provides
the access to querying for the state of the process and blocking on
completion. Additionally, it exposes the original ignoreExitValue flag
from the process spec. Finally the ExecHandleState internal enum has
been pulled up to public and renamed to ProcessHandleState.